fix: 修复编译错误（方法名、分页参数、类型转换等）

+++ lib/pages/my_page.dart
@@ -72,7 +72,7 @@ class _MyPageState extends State<MyPage> {
       // 新增：待审批代码修改数量
       if (results[2] != null && results[2] is int) {
         setState(() {
-          _pendingCodeFixCount = results[2];
+          _pendingCodeFixCount = results[2] as int;
         });
       }
     } catch (e) {
